fix(mcp): move MCPToolResultError to exceptions.py to survive utils reload
test_mcp_server_identity_env.py reloads mcp_server/utils.py to re-read its
env-derived constants, which forks any class defined there into two
identities; an MCPToolResultError instance created by server.py (holding
the pre-reload class) then fails isinstance checks against the post-reload
class, making test_fire_mcp_tool_call_logging_iserror_logs_failure
order-dependent (seen on the proxy-infra CI shard). exceptions.py is never
reloaded, so the class identity stays stable
